get products from search

+++ Salesman/app/src/main/java/cmpe295/sjsu/edu/salesman/fragments/OfferFragment.java
@@ -109,7 +109,7 @@ public class OfferFragment extends Fragment implements OfferCardAdapter.OnItemCl
                         }
                         */
 
-                    offers.add(new Offer(offerResponse.getProduct().getName(), Float.toString(offerResponse.getProduct().getPrice()), Color.parseColor("#d32f2f"), offerResponse.getProduct().getImage()));
+                    offers.add(new Offer(offerResponse.getProduct().getName(), Double.toString(offerResponse.getProduct().getPrice()), Color.parseColor("#d32f2f"), offerResponse.getProduct().getImage()));
                 }
 
                 // Note that results are not delivered on UI thread.
